Close #27

+++ ShoppingPoint/src/main/java/com/example/shoppingpoint/controller/PaymentController.java
@@ -102,7 +102,7 @@ public class PaymentController {
     }
 
     public int calculatePointsUsed(float fullTotal, int cardPoints, int pointsInEuro) {
-        float total = fullTotal - (float) (cardPoints / pointsInEuro);
+        float total = (float) (fullTotal - ((double) cardPoints / pointsInEuro));
         if (total < 0) return (int) Math.floor(fullTotal * pointsInEuro);
         return cardPoints;
     }
